Fixed missing init in tests
